# Conflicts:
#	web/pages/device/threshold.html
#	web/pages/output/relay_ctrl.html

+++ cores/outputs/op_zctrl.h
@@ -14,7 +14,7 @@ private:
     void writeCtrlCmd(uchar *cmd, int k);
     void funSwitch(uchar *on, uchar *off, uchar all=0) override;
     void funClearEle(uchar *buf) override;
-    void funDelay(uchar *buf) override;
+    void funDelay(uint *buf) override;
 };
 
 #endif // OP_ZCTRL_H
